{{cookiecutter.project_slug}}/requirements-dev.txt
django-livereload>=1.7
flake8>=3.0.4
pylint>=1.6.4
pylint-django>=0.7.2
pytest>=3.0.7
pytest-django>=3.1.2
